refactor: break up task_creation into managable sections
