+++ igneous/tasks/__init__.py
@@ -1,9 +1,10 @@
 from .skeletonization import SkeletonTask, UnshardedSkeletonMergeTask, ShardedSkeletonMergeTask
 from .mesh import MeshTask, MeshManifestTask, GrapheneMeshTask
 from .tasks import (
-  HyperSquareConsensusTask, 
+  HyperSquareConsensusTask, #HyperSquareTask,
   DownsampleTask, QuantizeTask, 
   TransferTask, WatershedRemapTask, DeleteTask, 
   LuminanceLevelsTask, ContrastNormalizationTask,
-  MaskAffinitymapTask, InferenceTask, BlackoutTask
+  MaskAffinitymapTask, InferenceTask, BlackoutTask,
+  TouchTask
 )
